Finalize barcode scanning logic
